Modified File socket program
The server delivers a file over a socket. In case of an error, informs that to the client as well.
If the file exists in the server, delivers the file onto the client. Client makes a new file named as $filename+"-copy". For example, if the file on server is called "hi", the client gets "hi-copy".

An example file called hi is available for testing purposes.